refactor: переместить тесты для валидации в пакет dto
